demo-phidata/requirements.txt
phidata
openai
duckduckgo-search
yfinance
newspaper4k
python-dotenv
lancedb
sqlalchemy